// src/types/taxonomy.ts
export type Tag = {
  id: string;
  name: string;
  color?: string | null;
  createdAt: string;
  updatedAt: string;
  deletedAt?: string | null;
};

export type Category = {
  id: string;
  name: string;
  color?: string | null;
  createdAt: string;
  updatedAt: string;
  deletedAt?: string | null;
};